// src/model/interface_tingkat_pengurus.ts
export interface ModelTingkatPengurus {
  id: string
    userId: string
    masterStatusKeanggotaanId: number
    masterTingkatPengurusId: number
    masterJabatanId: any
    masterJabatanDewanPembinaId: number
    masterJabatanDewanPimpinanPusatId: number
    masterJabatanDewanPimpinanDaerahId: number
    masterJabatanDewanPimpinanCabangId: number
    masterJabatanPimpinanAnakCabangId: number
    masterJabatanPimpinanRantingId: number
    masterJabatanPerwakilanPartaiDiLuarNegeriId: number
    masterSayapPartaiId: number
    masterKaderPartaiId: number
    masterProvinceId: number
    masterKabKotId: number
    masterKecamatanId: number
    masterDesaId: number
    masterNegaraId: number
    alamatKantor: string
    waAdmin: string
  
}